Move some parts of domainData to siloDomainData
Redux-Sync is now used to sync states between tabs, however
`siloDomainData` is skipped during sync. This allows browser tab
specific UI states while other data are kept in sync.
