+++ frontend/src/common/action-creators/siloDomainData.js
@@ -1,0 +1,15 @@
+import {
+    SET_ACTIVE_PROJECT,
+    SET_ACTIVE_COUNTRY,
+} from '../action-types/siloDomainData';
+
+
+export const setActiveProjectAction = ({ activeProject }) => ({
+    type: SET_ACTIVE_PROJECT,
+    activeProject,
+});
+
+export const setActiveCountryAction = ({ activeCountry }) => ({
+    type: SET_ACTIVE_COUNTRY,
+    activeCountry,
+});
