+++ frontend/src/common/action-types/domainData.js
@@ -9,11 +9,7 @@ export const SET_USER_GROUPS = 'domain-data/SET_USER_GROUPS';
 export const SET_USER_GROUP = 'domain-data/SET_USER_GROUP';
 export const UNSET_USER_GROUP = 'domain-data/UNSET_USER_GROUP';
 
-
-export const SET_ACTIVE_PROJECT = 'domain-data/SET_ACTIVE_PROJECT';
-export const SET_ACTIVE_COUNTRY = 'domain-data/SET_ACTIVE_COUNTRY';
 export const UNSET_REGION = 'domain-data/UNSET_REGION';
-
 export const ADD_NEW_COUNTRY = 'domain-data/ADD_NEW_COUNTRY';
 
 export const SET_LEADS = 'domain-data/SET_LEADS';
