+++ frontend/src/common/action-types/siloDomainData.js
@@ -1,0 +1,2 @@
+export const SET_ACTIVE_PROJECT = 'domain-data/SET_ACTIVE_PROJECT';
+export const SET_ACTIVE_COUNTRY = 'domain-data/SET_ACTIVE_COUNTRY';
