+++ frontend/src/common/initial-state/siloDomainData.js
@@ -1,0 +1,6 @@
+const initialSiloDomainData = {
+    activeProject: 1,
+    activeCountry: undefined,
+};
+
+export default initialSiloDomainData;
