+++ frontend/src/common/reducers/index.js
@@ -1,6 +1,7 @@
 import { combineReducers } from 'redux';
 import authReducer from './auth';
 import domainDataReducer from './domainData';
+import siloDomainDataReducer from './siloDomainData';
 import datetimeReducer from './datetime';
 import navbarReducer from './navbar';
 
@@ -8,6 +9,7 @@ const appReducer = combineReducers({
     auth: authReducer,
     domainData: domainDataReducer,
     datetime: datetimeReducer,
+    siloDomainData: siloDomainDataReducer,
     navbar: navbarReducer,
 });
