+++ frontend/src/common/redux.js
@@ -1,8 +1,10 @@
 export * from './action-creators/auth';
 export * from './action-creators/domainData';
+export * from './action-creators/siloDomainData';
 export * from './action-creators/navbar';
 
 export * from './selectors/auth';
 export * from './selectors/datetime';
 export * from './selectors/domainData';
+export * from './selectors/siloDomainData';
 export * from './selectors/navbar';
